+++ HISTORY.rst
@@ -61,6 +61,7 @@
 - Add the `json_load` templating filter.
 - Add a Docker Compose service to mount a cabinet.
 - Support disabling the search app.
+- Split the cabinet app views module.
 
 4.8.3 (2024-10-04)
 ==================